@@ -20,7 +20,6 @@ exports.isAuthenticated = catchAsyncErrors(async (req, res, next) => {
 
 exports.isSeller = catchAsyncErrors(async (req, res, next) => {
   const { seller_token } = req.cookies;
-
   if (!seller_token) {
     return next(new ErrorHandler("Please login to continue", 401));
   }
@@ -31,3 +30,14 @@ exports.isSeller = catchAsyncErrors(async (req, res, next) => {
 
   next();
 });
+
+exports.isAdmin = (...roles) => {
+  return (req, res, next) => {
+    if (!roles.includes(req.user.role)) {
+      return next(
+        new ErrorHandler(`${req.user.role} can not access this resources!`)
+      );
+    }
+    next();
+  };
+};
