@@ -6,7 +6,8 @@ const sendToken = (user, statusCode, res) => {
   const options = {
     expires: new Date(Date.now() + 90 * 24 * 60 * 60 * 1000),
     httpOnly: true,
-    Secure: true,
+    sameSite: "none",
+    secure: true,
   };
 
   res.status(statusCode).cookie("token", token, options).json({
