@@ -8,7 +8,7 @@ const Navbar = ({ active }) => {
     <div className={`block 800px:${styles.noramlFlex}`}>
       {navItems &&
         navItems.map((i, index) => (
-          <div key={index} className="flex">
+          <div className="flex">
             <Link
               to={i.url}
               className={`${
