@@ -1,16 +1,17 @@
 import React, { useEffect, useState } from "react";
+import { useSelector } from "react-redux";
 import { productData } from "../../static/data";
 import styles from "../../styles/styles";
 import ProductCard from "../Route/ProductCard/ProductCard";
 
 const SuggestedProduct = ({ data }) => {
-  const [products, setProducts] = useState(null);
+  const { allProducts } = useSelector((state) => state.products);
+  const [productData, setProductData] = useState();
 
   useEffect(() => {
     const d =
-      productData && productData.filter((i) => i.category === data.category);
-    setProducts(d);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
+      allProducts && allProducts.filter((i) => i.category === data.category);
+    setProductData(d);
   }, []);
 
   return (
@@ -23,8 +24,10 @@ const SuggestedProduct = ({ data }) => {
             Related Product
           </h2>
           <div className="grid grid-cols-1 gap-[20px] md:grid-cols-2 md:gap-[25px] lg:grid-cols-4 lg:gap-[25px] xl:grid-cols-5 xl:gap-[30px] mb-12">
-            {products &&
-              products.map((i, index) => <ProductCard data={i} key={index} />)}
+            {productData &&
+              productData.map((i, index) => (
+                <ProductCard data={i} key={index} />
+              ))}
           </div>
         </div>
       ) : null}
