@@ -1,9 +1,11 @@
 import React from "react";
-import { productData } from "../../../static/data";
+import { useSelector } from "react-redux";
 import styles from "../../../styles/styles";
 import ProductCard from "../ProductCard/ProductCard";
 
 const FeaturedProduct = () => {
+  const { allProducts } = useSelector((state) => state.products);
+
   return (
     <div>
       <div className={`${styles.section}`}>
@@ -11,8 +13,14 @@ const FeaturedProduct = () => {
           <h1>Featured Products</h1>
         </div>
         <div className="grid grid-cols-1 gap-[20px] md:grid-cols-2 md:gap-[25px] lg:grid-cols-4 lg:gap-[25px] xl:grid-cols-5 xl:gap-[30px] mb-12 border-0">
-          {productData &&
-            productData.map((i, index) => <ProductCard data={i} key={index} />)}
+          {allProducts && allProducts.length !== 0 && (
+            <>
+              {allProducts &&
+                allProducts.map((i, index) => (
+                  <ProductCard data={i} key={index} />
+                ))}
+            </>
+          )}
         </div>
       </div>
     </div>
