@@ -1,12 +1,13 @@
 import { React, useState } from "react";
 import { AiOutlineEye, AiOutlineEyeInvisible } from "react-icons/ai";
 import styles from "../../styles/styles";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import axios from "axios";
 import { server } from "../../server";
 import { toast } from "react-toastify";
 
 const ShopLogin = () => {
+  const navigate = useNavigate();
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [visible, setVisible] = useState(false);
@@ -25,6 +26,7 @@ const ShopLogin = () => {
       )
       .then((res) => {
         toast.success("Login Success!");
+        navigate("/dashboard");
         window.location.reload(true);
       })
       .catch((err) => {
