@@ -1,29 +1,40 @@
 import React, { useEffect, useState } from "react";
+import { useSelector } from "react-redux";
 import Header from "../components/Layout/Header";
+import Loader from "../components/Layout/Loader";
 import ProductCard from "../components/Route/ProductCard/ProductCard";
-import { productData } from "../static/data";
 import styles from "../styles/styles";
+import Footer from "../components/Layout/Footer";
 
 const BestSellingPage = () => {
   const [data, setData] = useState([]);
+  const { allProducts, isLoading } = useSelector((state) => state.products);
 
   useEffect(() => {
-    const d =
-      productData && productData.sort((a, b) => b.total_sell - a.total_sell);
-    setData(d);
-  }, []);
+    const allProductsData = allProducts ? [...allProducts] : [];
+    const sortedData = allProductsData?.sort((a, b) => b.sold_out - a.sold_out);
+    setData(sortedData);
+  }, [allProducts]);
 
   return (
-    <div>
-      <Header activeHeading={2} />
-      <br />
-      <br />
-      <div className={`${styles.section}`}>
-        <div className="grid grid-cols-1 gap-[20px] md:grid-cols-2 md:gap-[25px] lg:grid-cols-4 lg:gap-[25px] xl:grid-cols-5 xl:gap-[30px] mb-12">
-          {data && data.map((i, index) => <ProductCard data={i} key={index} />)}
+    <>
+      {isLoading ? (
+        <Loader />
+      ) : (
+        <div>
+          <Header activeHeading={2} />
+          <br />
+          <br />
+          <div className={`${styles.section}`}>
+            <div className="grid grid-cols-1 gap-[20px] md:grid-cols-2 md:gap-[25px] lg:grid-cols-4 lg:gap-[25px] xl:grid-cols-5 xl:gap-[30px] mb-12">
+              {data &&
+                data.map((i, index) => <ProductCard data={i} key={index} />)}
+            </div>
+          </div>
+          <Footer />
         </div>
-      </div>
-    </div>
+      )}
+    </>
   );
 };
 
