@@ -1,27 +1,35 @@
 import React, { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
+import { useParams, useSearchParams } from "react-router-dom";
 import Footer from "../components/Layout/Footer";
 import Header from "../components/Layout/Header";
 import ProductDetails from "../components/Products/ProductDetails";
-import { productData } from "../static/data";
 import SuggestedProduct from "../components/Products/SuggestedProduct";
+import { useSelector } from "react-redux";
 
 const ProductDetailsPage = () => {
-  const { name } = useParams();
+  const { allProducts } = useSelector((state) => state.products);
+  const { allEvents } = useSelector((state) => state.events);
+  const { id } = useParams();
   const [data, setData] = useState(null);
-  const productName = name.replace(/-/g, " ");
+  const [searchParams] = useSearchParams();
+  const eventData = searchParams.get("isEvent");
 
   useEffect(() => {
-    const data = productData.find((i) => i.name === productName);
-    setData(data);
+    if (eventData !== null) {
+      const data = allEvents && allEvents.find((i) => i._id === id);
+      setData(data);
+    } else {
+      const data = allProducts && allProducts.find((i) => i._id === id);
+      setData(data);
+    }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [allProducts, allEvents]);
 
   return (
     <div>
       <Header />
       <ProductDetails data={data} />
-      {data && <SuggestedProduct data={data} />}
+      {!eventData && <>{data && <SuggestedProduct data={data} />}</>}
       <Footer />
     </div>
   );
