@@ -5,14 +5,14 @@ import ShopLogin from "../components/Shop/ShopLogin";
 
 const ShopLoginPage = () => {
   const navigate = useNavigate();
-  const { isSeller, seller } = useSelector((state) => state.seller);
+  const { isSeller, isLoading } = useSelector((state) => state.seller);
 
   useEffect(() => {
     if (isSeller === true) {
-      navigate(`/shop/${seller._id}`);
+      navigate(`/dashboard`);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [isLoading, isSeller]);
   return (
     <div>
       <ShopLogin />
