@@ -1,6 +1,6 @@
-import React, { useEffect } from "react";
-import { useSelector } from "react-redux";
-import { useNavigate } from "react-router-dom";
+import React, { useEffect } from 'react'
+import { useSelector } from 'react-redux';
+import { useNavigate } from 'react-router-dom';
 import Signup from "../components/Signup/Signup";
 
 const SignupPage = () => {
@@ -8,16 +8,15 @@ const SignupPage = () => {
   const { isAuthenticated } = useSelector((state) => state.user);
 
   useEffect(() => {
-    if (isAuthenticated === true) {
+    if(isAuthenticated === true){
       navigate("/");
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [])
   return (
     <div>
-      <Signup />
+        <Signup />
     </div>
-  );
-};
+  )
+}
 
-export default SignupPage;
+export default SignupPage
