@@ -1,7 +1,7 @@
 import { createReducer } from "@reduxjs/toolkit";
 
 const initialState = {
-  isSeller: false,
+  isLoading: true,
 };
 
 export const sellerReducer = createReducer(initialState, {
@@ -17,6 +17,19 @@ export const sellerReducer = createReducer(initialState, {
     state.isLoading = false;
     state.error = action.payload;
     state.isSeller = false;
+  },
+
+  // get all sellers ---admin
+  getAllSellersRequest: (state) => {
+    state.isLoading = true;
+  },
+  getAllSellersSuccess: (state, action) => {
+    state.isLoading = false;
+    state.sellers = action.payload;
+  },
+  getAllSellerFailed: (state, action) => {
+    state.isLoading = false;
+    state.error = action.payload;
   },
   clearErrors: (state) => {
     state.error = null;
